requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.7.4.post0
    # via -r requirements.in
async-timeout==3.0.1
    # via aiohttp
attrs==21.2.0
    # via aiohttp
bidict==0.21.2
    # via python-socketio
certifi==2021.5.30
    # via requests
chardet==4.0.0
    # via
    #   aiohttp
    #   requests
cycler==0.10.0
    # via matplotlib
dnspython==1.16.0
    # via eventlet
eventlet==0.31.0
    # via -r requirements.in
greenlet==1.1.0
    # via eventlet
idna==2.10
    # via
    #   requests
    #   yarl
kiwisolver==1.3.1
    # via matplotlib
matplotlib==3.4.2
    # via
    #   -r requirements.in
    #   seaborn
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
numpy==1.21.0
    # via
    #   -r requirements.in
    #   matplotlib
    #   pandas
    #   scipy
    #   seaborn
pandas==1.2.5
    # via seaborn
pillow==8.2.0
    # via matplotlib
pyparsing==2.4.7
    # via matplotlib
python-dateutil==2.8.1
    # via
    #   matplotlib
    #   pandas
python-engineio==3.13.1
    # via python-socketio
python-socketio==4.6.0
    # via -r requirements.in
pytz==2021.1
    # via pandas
requests==2.25.1
    # via -r requirements.in
scipy==1.7.0
    # via seaborn
seaborn==0.11.1
    # via -r requirements.in
six==1.16.0
    # via
    #   cycler
    #   eventlet
    #   python-dateutil
typing-extensions==3.10.0.0
    # via aiohttp
urllib3==1.26.6
    # via requests
yarl==1.6.3
    # via aiohttp
